Added mocha testing and istanbul coverage
