+++ lib/index.js
@@ -1,5 +1,7 @@
 'use strict'
 
+var routes = require("./routes")
+
 const Hapi = require('hapi')
 
 // Create a server with a host and port
@@ -10,23 +12,10 @@ server.connection({
 })
 
 // Add the route
-server.route({
-  method: 'GET',
-  path:   '/version',
-  config: {
-    state: {
-      parse:      true,
-      failAction: 'log'
-    }
-  },
-  handler: function (request, reply) {
-    return reply({ version: process.env.npm_package_version })
-  }
-})
+server.route(routes)
 
 // Start the server
 server.start((err) => {
-
   if (err) {
     throw err
   }
